Rename SelectWithLabelProps
